Set the GuildID on the member object when it was fetched
When I was writing the GetPermissions function on the client I completely forgot about this caveat, causing it to be broken from the Client struct (sorry - that's my bad, I shoud've tested before I PR'd). This fixes that and also makes it easier to access the guild ID with a member object since it doesn't need to be patched in.