Add IsByWebhook and NotByWebhook Message Filters.
This commit adds additional default message filters for determining if a message is sourced from a Discord Webhook. Currently, IsByBot and NotByBot exist, but these will consider webhooks "not a bot," as bot will be set to false.

This leverages webhookid on messages being set or unset to determine if the message is sent by a webhook. This method is consistent with the public API docs, which recommend exactly this:

> If the message is generated by a webhook, the author object corresponds to the webhook's id, username, and avatar. You can tell if a message is generated by a webhook by checking for the webhook_id on the message object.

Source:
https://discord.com/developers/docs/resources/channel#message-object-message-structure